% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_embeddings.R
\name{align_embeddings}
\alias{align_embeddings}
\title{Align Two Embeddings Models}
\usage{
align_embeddings(x, y, matching = NULL)
}
\arguments{
\item{x}{an embeddings object}

\item{y}{an embeddings object. If \code{matching = NULL}, \code{y} must contain at least
a few rownames matching those of \code{x}.}

\item{matching}{(optional) a named character vector specifying a one-to-one
matching between rownames of \code{x} (names) and rownames of \code{y} (values)}
}
\description{
Rotates the embeddings in \code{x} so that they can be compared with those in \code{y}.
}
\details{
Computes orthogonal Procrustes as described by Schönemann (1966). This
computation is based only on rows with names that appear in both \code{x} and \code{y}
(or on rows specified by \code{matching}), but the output includes all rows of \code{x}.
}
\section{Value}{

An embeddings object with the same rownames and number of rows as \code{x}, rotated
(and with reduced or increased dimensionality, if need be) to align with \code{y}.
}

\references{
Schönemann, P. H. (1966). A generalized solution of the orthogonal
procrustes problem. Psychometrika, 31(1), 1–10. https://doi.org/10.1007/BF02289451
}
