% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{as.embeddings}
\alias{as.embeddings}
\title{Coercion and checking functions for embeddings objects}
\usage{
as.embeddings(x)
}
\arguments{
\item{mat}{a matrix}
}
\description{
An embeddings object is a numeric matrix with rownames (generally tokens). Convert an
eligible input object into an embeddings object, or check whether an object
is an embeddings object.
}
\examples{
   random_mat <- matrix(
      sample(1:10, 20, replace = TRUE),
      nrow = 2,
      dimnames = list(c("happy", "sad"))
      )
   as.embeddings(random_mat)
}
