% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{build_token_index}
\alias{build_token_index}
\title{Build hash table to map tokens (i.e. doc_ids) to matrix indices}
\usage{
build_token_index(embeddings)
}
\arguments{
\item{embeddings}{an embeddings object}
}
\description{
Build hash table to map tokens (i.e. doc_ids) to matrix indices
}
\keyword{internal}
