% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{dot_prod}
\alias{dot_prod}
\title{Similarity and Distance Metrics}
\usage{
dot_prod(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector the same length as x}

\item{pos}{a numeric vector the same length as x; the positive end of the anchored vector}

\item{neg}{a numeric vector the same length as x; the negative end of the anchored vector}
}
\description{
Metrics for measuring relationships between vector embeddings.
}
\details{
`dot_prod` gives the dot product. `cos_sim` gives the cosine similarity (i.e.
the dot product of two normalized vectors). `euc_dist` gives the Euclidean
distance. `anchored_sim` gives the position of `x` on the spectrum between two
anchor points, where vectors aligned with `pos` are given a score of 1 and those
aligned with `neg` are given a score of 0. For more on anchored vectors, see
[Data Science for Psychology: Natural Language, Chapter 20](https://ds4psych.com/navigating-vectorspace#sec-dimension-projection).
}
\examples{
   vec1 <- c(1, 5, 2)
   vec2 <- c(4, 2, 2)
   vec3 <- c(1, -2, -13)

   dot_prod(vec1, vec2)
   cos_sim(vec1, vec2)
   euc_dist(vec1, vec2)
   anchored_sim(vec1, vec2, vec3)
}
