% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_tokens.R
\name{embed_tokens}
\alias{embed_tokens}
\alias{embed_tokens.default}
\alias{embed_tokens.data.frame}
\title{Get Embeddings of Tokens in a Text or Corpus}
\usage{
embed_tokens(x, ...)

\method{embed_tokens}{default}(
  x,
  model,
  ...,
  .keep_missing = FALSE,
  tolower = TRUE,
  output_embeddings = FALSE
)

\method{embed_tokens}{data.frame}(
  x,
  text_col,
  model,
  id_col = NULL,
  ...,
  .keep_missing = FALSE,
  .keep_all = FALSE,
  tolower = TRUE,
  output_embeddings = FALSE
)
}
\arguments{
\item{x}{a character vector of texts, a data frame, a 'quanteda'
\link[quanteda:tokens]{tokens} object, or a 'quanteda' \link[quanteda:corpus]{corpus}}

\item{...}{additional parameters to pass to \code{quanteda::tokens()}}

\item{model}{a word embeddings model in the form of an \link{embeddings} object}

\item{.keep_missing}{logical. What should be done about tokens in \code{x} that
are not present in \code{model}? If \code{FALSE} (the default), they will be ignored.
If \code{TRUE}, they will be returned as \code{NA}.}

\item{tolower}{logical. Convert all text to lowercase?}

\item{output_embeddings}{\code{FALSE} (the default) returns a tibble.
\code{TRUE} returns a list of embeddings objects. See 'Value' for details.}

\item{text_col}{string. a column of texts to be tokenized and converted to embeddings}

\item{id_col}{optional string. column of unique document ids}

\item{.keep_all}{logical. Keep all columns from input? Ignored if \code{output_embeddings = TRUE}.}
}
\description{
Given a character vector or 'quanteda' object (\link[quanteda:tokens]{tokens}, \link[quanteda:dfm]{dfm}, or \link[quanteda:corpus]{corpus})  and a
word embeddings model in the form of an \link{embeddings} object, \code{\link[=embed_tokens]{embed_tokens()}}
returns the embedding for each token.
}
\section{Value}{

If \code{output_embeddings = FALSE}, a tibble with columns \code{doc_id}, \code{token}, and embedding
dimension names. If \code{.keep_all = TRUE}, the new columns will appear after
existing ones, and the class of the input will be maintained.
If \code{output_embeddings = TRUE}, a named list of embeddings objects with tokens as
rownames.
}

\examples{
texts <- c("this says one thing", "and this says another")
texts_token_embeddings <- embed_tokens(texts, glove_twitter_25d)
texts_token_embeddings

# quanteda workflow
library(quanteda)
texts_tokens <- tokens(texts)
texts_token_embeddings <- embed_tokens(texts_tokens, glove_twitter_25d)
texts_token_embeddings

# dplyr workflow
texts_df <- data.frame(text = texts)
texts_token_embeddings <- texts_df |> embed_tokens("text", glove_twitter_25d)
texts_token_embeddings
}
\seealso{
\code{\link[=embed_docs]{embed_docs()}}
}
