% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings-methods.R
\name{rownames<-.embeddings}
\alias{rownames<-.embeddings}
\alias{dimnames<-.embeddings}
\alias{unique.embeddings}
\alias{t.embeddings}
\alias{rbind.embeddings}
\alias{as.matrix.embeddings}
\alias{as_tibble.embeddings}
\title{Methods for Embeddings Objects}
\usage{
\method{rownames}{embeddings}(x) <- value

\method{dimnames}{embeddings}(x) <- value

\method{unique}{embeddings}(x, incomparables = FALSE, MARGIN = 1, fromLast = FALSE, ...)

\method{t}{embeddings}(x)

\method{rbind}{embeddings}(..., deparse.level = 1)

\method{as.matrix}{embeddings}(x, ...)

\method{as_tibble}{embeddings}(x, ..., rownames = NULL)
}
\arguments{
\item{x}{an embeddings object}

\item{value}{a valid value for that component of
    \code{\link[base]{dimnames}(x)}.  For a matrix or array this is either
    \code{NULL} or a character vector of non-zero length equal to the
    appropriate dimension.}

\item{incomparables}{a vector of values that cannot be compared.
    \code{FALSE} is a special value, meaning that all values can be
    compared, and may be the only value accepted for methods other than
    the default.  It will be coerced internally to the same type as
       \code{x}.}

\item{MARGIN}{the array margin to be held fixed: a single integer.}

\item{fromLast}{logical indicating if duplication should be considered
    from the last, i.e., the last (or rightmost) of identical elements will
    be kept.  This only matters for \code{\link[base]{names}} or
    \code{\link[base]{dimnames}}.}

\item{...}{for \code{rbind}, embeddings objects or other objects that will be coerced
to embeddings. Otherwise, arguments for particular methods.}

\item{deparse.level}{integer controlling the construction of labels in
    the case of non-matrix-like arguments (for the default method):\cr
    \code{deparse.level = 0} constructs no labels;\cr
    the default \code{deparse.level = 1} typically and
    \code{deparse.level = 2} always construct labels from the argument
    names, see the \sQuote{Value} section below.}

\item{rownames}{How to treat existing row names of a data frame or matrix:
\itemize{
\item \code{NULL}: remove row names. This is the default.
\item \code{NA}: keep row names.
\item A string: the name of a new column. Existing rownames are transferred
into this column and the \code{row.names} attribute is deleted.
No name repair is applied to the new column name, even if \code{x} already contains
a column of that name.
Use \code{as_tibble(rownames_to_column(...))} to safeguard against this case.
}

Read more in \link[tibble]{rownames}.}
}
\description{
Functions that modify the rownames or indices of embeddings objects
(e.g. \code{unique}, \code{rbind}, \verb{rownames<-} etc.) will behave as usual
for matrices while automatically updating the hash table as necessary.
Functions that coerce embeddings objects to other data types (e.g. \code{as.matrix})
will strip the hash table attribute, leaving a clean, familiar object.
}
