% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{as.embeddings}
\alias{as.embeddings}
\alias{as.embeddings.data.frame}
\alias{is.embeddings}
\alias{as.matrix.embeddings}
\title{Coercion and checking functions for embeddings objects}
\usage{
as.embeddings(x, ...)

\method{as.embeddings}{data.frame}(x, id_col = NULL, ..., .rowname_repair = TRUE)

is.embeddings(x, ...)

\method{as.matrix}{embeddings}(x, ...)
}
\arguments{
\item{x}{A data frame to be converted into embeddings.}

\item{...}{Additional arguments passed to or from other methods.}

\item{id_col}{Optional name of a column to take row names from.}

\item{.rowname_repair}{logical. If \code{TRUE} (the default), check that unique
rownames are provided, and name rows "doc_1", "doc_2", etc. if not.}
}
\description{
An embeddings object is a numeric matrix with rownames (generally tokens). Convert an
eligible input object into an embeddings object, or check whether an object
is an embeddings object.
}
\examples{
random_mat <- matrix(
  sample(1:10, 20, replace = TRUE),
  nrow = 2,
  dimnames = list(c("happy", "sad"))
  )
random_embeddings <- as.embeddings(random_mat)
as.matrix(random_embeddings)

tibble::as_tibble(random_embeddings, rownames = "token")
}
