% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{embeddings_check}
\alias{embeddings_check}
\title{Check if and object can be coerced to embeddings}
\usage{
embeddings_check(x)
}
\arguments{
\item{x}{an object to be checked for coercability to embeddings}
}
\description{
Check if and object can be coerced to embeddings
}
\keyword{internal}
