% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest.R
\name{find_nearest}
\alias{find_nearest}
\title{Find Nearest Tokens in Embedding Space}
\usage{
find_nearest(
  object,
  newdata,
  top_n = 10L,
  method = c("cosine", "euclidean", "minkowski", "dot_prod", "anchored"),
  ...,
  each = FALSE,
  include_self = TRUE,
  decreasing = NULL
)
}
\arguments{
\item{object}{an embeddings object made by \code{load_embeddings()} or \code{as.embeddings()}}

\item{newdata}{a character vector of tokens indexed by \code{object},
an embeddings object of the same dimensionality as \code{object},
or a numeric vector of the same dimensionality as \code{object}.}

\item{top_n}{integer. How many nearest neighbors should be output? If \code{each = TRUE},
how many nearest neighbors should be output for each token?}

\item{method}{either the name of a method to compute similarity or distance,
or a function that takes two vectors and outputs a scalar, like those listed
in \link[=sim_metrics]{Similarity and Distance Metrics}. The value is passed to \code{\link[=get_similarities]{get_similarities()}}.}

\item{...}{additional parameters to be passed to method function by way of \code{\link[=get_similarities]{get_similarities()}}}

\item{each}{logical. If \code{FALSE} (the default), the embeddings of newdata are
averaged and the function will output an embeddings object with the \code{top_n}
nearest tokens to the overall average embedding. If \code{TRUE}, the function will
output a named list with one embeddings object for each token in newdata.}

\item{include_self}{logical. Should the token(s) in \code{newdata} be included in
the output?}

\item{decreasing}{logical. If \code{TRUE} (the default for \code{"cosine"}, \code{"dot_prod"},
and \code{"anchored"} methods), the embeddings with the highest similarity scores are
output. If \code{FALSE} (the default for \code{"euclidean"} and \code{"minkowski"} methods),
the embeddings with the lowest distance scores are output.}
}
\description{
Find Nearest Tokens in Embedding Space
}
\section{Value}{

An embeddings object or list of embeddings objects.
}

\examples{
words <- c("happy", "sad")

find_nearest(glove_twitter_25d, words)
find_nearest(glove_twitter_25d, words, each = TRUE)
}
