% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings-methods.R
\name{format.embeddings}
\alias{format.embeddings}
\title{Formatting and Printing for Embeddings Objects}
\usage{
\method{format}{embeddings}(x, ..., n = getOption("max.print"), round = 2)
}
\arguments{
\item{x}{the embeddings object to be formatted}

\item{...}{arguments to be passed to methods}

\item{n}{integer. How many rows should be printed? Defaults to \code{getOption("max.print")}}

\item{round}{integer. the number of decimal places to be displayed}
}
\description{
Since the numerical content of embeddings is generally not very informative
to the naked eye, the default formatting of matrices is often cumbersome.
As such, embeddings objects print in a format that allows quick visualization
of dimensionality and row names.
}
