% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glove_twitter_100d}
\alias{glove_twitter_100d}
\title{100-d GloVe Model Pretrained on Twitter (subset)}
\format{
An object of class \code{embeddings} (inherits from \code{matrix}, \code{array}) with 149554 rows and 100 columns.
}
\usage{
glove_twitter_100d
}
\description{
An embeddings object containing a subset of glove.twitter.100d pretrained by
Pennington et al. (2014) on 2B Tweets. The subset includes only embeddings for
common English words, as defined by [Peter Norvig's compilation of the 1/3
million most frequent English words](https://norvig.com/ngrams/count_1w.txt),
derived from the Google Web Trillion Word Corpus.
Full pretrained GloVe models are available for download at https://nlp.stanford.edu/projects/glove/
}
\references{
Pennington, J., Socher, R., and Manning, C. D. (2014).
  GloVe: Global Vectors for Word Representation. https://nlp.stanford.edu/projects/glove/
}
\keyword{data}
\keyword{glove_twitter_100d}
