% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{is_matrixlike}
\alias{is_matrixlike}
\title{Check if an object is matrix-like and numeric}
\usage{
is_matrixlike(x)
}
\arguments{
\item{x}{an object to be checked}
}
\description{
Check if an object is matrix-like and numeric
}
\keyword{internal}
