% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embeddings.R
\name{load_embeddings}
\alias{load_embeddings}
\title{Load Embeddings Model From File}
\usage{
load_embeddings(
  path,
  dimensions = as.numeric(str_extract(path, "[:digit:]+(?=d\\\\.txt)"))
)
}
\arguments{
\item{path}{a file path}

\item{dimensions}{the dimensionality of the embeddings. The defaults assumes
that this information is included in the file name as per the standard for
GloVe models.}
}
\description{
Load Embeddings Model From File
}
\details{
The file must have tokens in the first column and numbers in all other columns.
}
\section{Value}{

An embeddings object (a numeric matrix with tokens as rownames)
}

