% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude}
\alias{magnitude}
\alias{magnitude.numeric}
\title{Magnitude of Embeddings}
\usage{
magnitude(x, ...)

\method{magnitude}{numeric}(x, ...)
}
\arguments{
\item{x}{a numeric vector or embeddings object}

\item{...}{additional parameters to be passed to methods}
}
\description{
The magnitude, norm, or length of a vector is its Euclidean distance from the origin.
}
\section{Value}{

a numeric vector with one item per embedding in \code{x}
}

\examples{
vec <- c(1, 4, 2)
magnitude(vec)

valence_embeddings <- predict(glove_twitter_25d, c("good", "bad"))
magnitude(valence_embeddings)
}
