% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize.numeric}
\alias{normalize_rows}
\alias{normalize_rows.default}
\alias{normalize_rows.data.frame}
\title{Scale Embeddings to the Unit Hypersphere}
\usage{
normalize(x, ...)

\method{normalize}{numeric}(x, ...)

normalize_rows(x, ...)

\method{normalize_rows}{default}(x, ...)

\method{normalize_rows}{data.frame}(x, cols, ...)
}
\arguments{
\item{x}{for \code{normalize()}, a numeric vector, embeddings object, or list of numeric or embeddings objects.
For \code{normalize_rows()}, a matrix or dataframe with one embedding per row.}

\item{...}{additional parameters to be passed to methods}

\item{cols}{tidyselect - columns that contain numeric embedding values}
}
\description{
Normalize embeddings such that their magnitude is 1, while their angle from the origin is unchanged.
}
\section{Value}{

An object of the same class and dimensionality as x
}

\examples{
vec <- c(1, 4, 2)
normalize(vec)

valence_embeddings <- emb(glove_twitter_25d, c("good", "bad"))
normalize(valence_embeddings)

embeddings_list <- find_nearest(glove_twitter_25d, c("good", "bad"), each = TRUE)
normalize(embeddings_list)

valence_df <- tibble::as_tibble(valence_embeddings, rownames = "token")
valence_df |> normalize_rows(dim_1:dim_25)
}
