% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embeddings.R
\name{read_embeddings}
\alias{read_embeddings}
\title{Read Embeddings From a Text or Binary File}
\usage{
read_embeddings(path, words = NULL)
}
\arguments{
\item{path}{a file path}

\item{words}{optional list of words for which to retrieve embeddings.}
}
\description{
Reads GloVe text format, fastText text format, word2vec binary format,
and most tabular formats (csv, tsv, etc.)
}
\details{
If using a custom tabular format, the file must have tokens in the first column and numbers in all other columns.
}
\section{Value}{

An embeddings object (a numeric matrix with tokens as rownames)
}

