% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensionality.R
\name{reduce_dimensionality}
\alias{reduce_dimensionality}
\title{Reduce the Dimensionality of Embeddings}
\usage{
reduce_dimensionality(x)
}
\arguments{
\item{data}{a dataframe with one embedding per row}

\item{cols}{tidyselect - columns that contain numeric embedding values}

\item{reduce_to}{number of dimensions to keep}

\item{scale}{perform scaling in addition to centering?}
}
\description{
Includes methods for dataframes (in the style of `dplyr`) embeddings objects, and matrices.
}
\details{
The file must have tokens in the first column and numbers in all other columns.
}
\section{Value}{

An embeddings object (a numeric matrix with tokens as rownames)
}

