% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_matrix.R
\name{sim_matrix}
\alias{sim_matrix}
\alias{sim_matrix.data.frame}
\title{Pairwise Similarity or Distance Matrix}
\usage{
sim_matrix(x, ...)

\method{sim_matrix}{data.frame}(
  x,
  cols,
  method = c("cosine", "cosine_squished", "euclidean", "minkowski", "dot_prod",
    "anchored"),
  ...,
  tidy_output = FALSE
)
}
\arguments{
\item{x}{an embeddings object, matrix, or dataframe with one embedding per row}

\item{...}{additional parameters to be passed to method function}

\item{cols}{tidyselect - columns that contain numeric embedding values}

\item{method}{either the name of a method to compute similarity or distance,
or a function that takes two vectors, \code{x} and \code{y}, and outputs a scalar,
similar to those listed in \link[=sim_metrics]{Similarity and Distance Metrics}}

\item{tidy_output}{logical. If \code{FALSE} (the default), output a \link[stats:dist]{stats::dist}
object. If \code{TRUE}, output a tibble with columns \code{doc_id_1}, \code{doc_id_2}, and
the similarity or distance metric.}
}
\description{
Calculate a matrix of similarity scores between the rows of the input.
}
\details{
\subsection{Available Methods}{

When \code{method} is the name of one of the following supported methods,
computations are done with matrix operations and are therefore blazing fast.
\itemize{
\item \code{cosine}: cosine similarity
\item \code{cosine_squished}: cosine similarity, rescaled to range from 0 to 1
\item \code{euclidean}: Euclidean distance
\item \code{minkowski}: Minkowski distance; requires parameter \code{p}. When \code{p = 1}
(the default), this is the Manhattan distance. When \code{p = 2}, it is the
Euclidean distance. When \code{p = Inf}, it is the Chebyshev distance.
\item \code{dot_prod}: Dot product
}
When \code{method} is a custom function, operations are performed for each row and
may be slow for large inputs.
}
}
\section{Value}{

If \code{tidy_output = FALSE} (the default), a \link[stats:dist]{stats::dist} object.
If \code{tidy_output = TRUE}, a tibble with columns \code{doc_id_1}, \code{doc_id_2}, and
the similarity or distance metric.
}

\examples{
emb <- predict(glove_twitter_25d, c("table", "chair", "cat"))

sim_matrix(emb)
sim_matrix(emb, method = "euclidean")
sim_matrix(emb, method = function(x, y) sum(abs(x - y)))

valence_df <- tibble::as_tibble(emb, rownames = "token")
valence_df |> sim_matrix(dim_1:dim_25, tidy_output = TRUE)
}
