% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{dot_prod}
\alias{dot_prod}
\alias{sim_metrics}
\alias{cos_sim}
\alias{euc_dist}
\alias{anchored_sim}
\title{Similarity and Distance Metrics}
\usage{
dot_prod(x, y)

cos_sim(x, y)

euc_dist(x, y)

anchored_sim(x, pos, neg)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector the same length as x}

\item{pos, neg}{a pair of numeric vectors the same length as x; the positive and negative ends of the anchored vector}
}
\description{
Metrics for measuring relationships between vector embeddings.
}
\details{
\code{dot_prod} gives the dot product. \code{cos_sim} gives the cosine similarity (i.e.
the dot product of two normalized vectors). \code{euc_dist} gives the Euclidean
distance. \code{anchored_sim} gives the position of \code{x} on the spectrum between two
anchor points, where vectors aligned with \code{pos} are given a score of 1 and those
aligned with \code{neg} are given a score of 0. For more on anchored vectors, see
\href{https://ds4psych.com/navigating-vectorspace#sec-dimension-projection}{Data Science for Psychology: Natural Language, Chapter 20}.
}
\examples{
vec1 <- c(1, 5, 2)
vec2 <- c(4, 2, 2)
vec3 <- c(1, -2, -13)

dot_prod(vec1, vec2)
cos_sim(vec1, vec2)
euc_dist(vec1, vec2)
anchored_sim(vec1, vec2, vec3)
}
