% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{dot_prod_mat_vec}
\alias{dot_prod_mat_vec}
\alias{cos_sim_mat_vec}
\alias{cos_sim_squished_mat_vec}
\alias{euc_dist_mat_vec}
\alias{minkowski_dist_mat_vec}
\alias{anchored_sim_mat_vec}
\title{Similarity and Distance Between A Matrix and a Vector}
\usage{
dot_prod_mat_vec(x, y)

cos_sim_mat_vec(x, y)

cos_sim_squished_mat_vec(x, y)

euc_dist_mat_vec(x, y)

minkowski_dist_mat_vec(x, y, p = 1)

anchored_sim_mat_vec(x, pos, neg)
}
\arguments{
\item{x}{a numeric matrix or embeddings object}

\item{y}{a numeric vector of length \code{ncol(x)}}

\item{p}{\href{https://en.wikipedia.org/wiki/Lp_space#The_p-norm_in_finite_dimensions}{p-norm} used to compute the Minkowski distance}
}
\description{
These functions are equivalent to calling \link[=sim_metrics]{vector similarity metrics}
on each row of a matrix, but use matrix operations and are therefore much faster.
}
\section{Value}{

A named numeric vector of length \code{nrow(x)}
}

\keyword{internal}
