% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{dot_prod_matrix}
\alias{dot_prod_matrix}
\alias{cos_sim_matrix}
\alias{cos_sim_squished_matrix}
\alias{euc_dist_matrix}
\alias{minkowski_dist_matrix}
\alias{anchored_sim_matrix}
\title{Faster Similarity and Distance Metrics With Matrix Operations}
\usage{
dot_prod_matrix(x, y)

cos_sim_matrix(x, y)

cos_sim_squished_matrix(x, y)

euc_dist_matrix(x, y)

minkowski_dist_matrix(x, y, p = 1)

anchored_sim_matrix(x, pos, neg)
}
\arguments{
\item{x}{a numeric matrix or embeddings object}

\item{y}{a numeric vector of length \code{ncol(x)}}
}
\description{
Faster Similarity and Distance Metrics With Matrix Operations
}
\section{Value}{

A named numeric vector of length \code{nrow(x)}
}

\keyword{internal}
