% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{dot_prod_parallel}
\alias{dot_prod_parallel}
\alias{cos_sim_parallel}
\alias{cos_sim_squished_parallel}
\alias{euc_dist_parallel}
\alias{minkowski_dist_parallel}
\title{Parallel Similarity and Distance}
\usage{
dot_prod_parallel(x, y)

cos_sim_parallel(x, y)

cos_sim_squished_parallel(x, y)

euc_dist_parallel(x, y)

minkowski_dist_parallel(x, y, p = 1)
}
\arguments{
\item{x}{a numeric matrix or embeddings object}

\item{y}{a numeric matrix or embeddings object with the same dimensions as \code{x}}

\item{p}{\href{https://en.wikipedia.org/wiki/Lp_space#The_p-norm_in_finite_dimensions}{p-norm} used to compute the Minkowski distance}
}
\description{
These functions compute parallel \link[=sim_metrics]{similarity metrics} between
each row of a matrix and its corresponding row in another matrix.
}
\section{Value}{

A named numeric vector of length \code{nrow(x)}
}

\keyword{internal}
