% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_embedding.R
\name{textstat_embedding}
\alias{textstat_embedding}
\alias{embed_docs}
\title{Get Text Embeddings by Averaging Word Embeddings}
\usage{
textstat_embedding(dfm, model, w = NULL, method = "mean", output = "tibble")

embed_docs(
  data,
  text_col,
  model,
  id_col = NULL,
  w = NULL,
  method = "mean",
  ...,
  .keep_all = FALSE,
  tolower = TRUE,
  output = "tibble"
)
}
\arguments{
\item{dfm}{a quanteda dfm}

\item{model}{an \link{embeddings} object. For \code{embed_docs()}, \code{model}
can alternatively be a function that takes a character vector and outputs a
dataframe with a row for each element of the input.}

\item{w}{optional weighting for embeddings in \code{model} if \code{model} is an
embeddings object. See \code{\link[=average_embedding]{average_embedding()}}.}

\item{method}{method to use for averaging. See \code{\link[=average_embedding]{average_embedding()}}. Note that
\code{method = "median"} does not use matrix operations and may therefore be slow
for datasets with many documents.}

\item{output}{\code{"tibble"} (the default) returns a tibble. \code{"embeddings"}
returns an embeddings object. See 'Value' for details.}

\item{data}{a data frame or data frame extension (e.g. a tibble)}

\item{text_col}{string. a column of texts for which to compute embeddings}

\item{id_col}{optional string. column of unique document ids}

\item{...}{additional parameters to pass to \code{quanteda::tokens()} or to the
user-specified modeling function}

\item{.keep_all}{logical. Keep all columns from input? Ignored if \code{output = "tibble"}.}

\item{tolower}{logical. Convert all text to lowercase? If \code{model} is an
embeddings object, this value is passed to \code{\link[quanteda:dfm]{quanteda::dfm()}}.}
}
\description{
\code{textstat_embedding()} takes a 'quanteda' dfm. \code{embed_docs()} is a more
versatile function for use in a purely dplyr-style workflow, which acts directly
on a column of texts.
}
\section{Value}{

If \code{output = "tibble"}, a tibble with columns \code{doc_id}, and \code{dim_1}, \code{dim_2},
etc. or similar. If \code{.keep_all = TRUE}, the new columns will appear after
existing ones.
If \code{output = "embeddings"}, an embeddings object with document ids as rownames.
}

\examples{
# quanteda workflow
library(quanteda)
texts <- c("this says one thing", "and this says another")
texts_dfm <- dfm(tokens(texts))

texts_embeddings <- textstat_embedding(texts_dfm, glove_twitter_25d)
texts_embeddings

# dplyr workflow
texts_df <- data.frame(text = texts)
texts_embeddings <- texts_df |> embed_docs("text", glove_twitter_25d)
texts_embeddings
}
