% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{token_index_add}
\alias{token_index_add}
\title{Update hash table after additions}
\usage{
token_index_add(embeddings, added_tokens)
}
\arguments{
\item{embeddings}{an embeddings object}

\item{added_tokens}{tokens to be added to the hash table}
}
\description{
Update hash table after additions
}
\keyword{internal}
