% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_dist.R
\name{total_dist}
\alias{total_dist}
\alias{average_sim}
\title{Compare Two Embedding Models}
\usage{
total_dist(
  x,
  y,
  matching = NULL,
  method = c("euclidean", "minkowski", "cosine", "cosine_squished", "dot_prod"),
  average = FALSE,
  ...
)

average_sim(x, y, matching = NULL, method = "cosine", average = TRUE, ...)
}
\arguments{
\item{x}{an embeddings object}

\item{y}{an embeddings object. If \code{matching = NULL}, \code{y} must contain at least
a few rownames matching those of \code{x}.}

\item{matching}{(optional) a named character vector specifying a one-to-one
matching between rownames of \code{x} (names) and rownames of \code{y} (values)}

\item{method}{either the name of a method to compute similarity or distance,
or a function that takes two vectors, \code{x} and \code{y}, and outputs a scalar,
similar to those listed in \link[=sim_metrics]{Similarity and Distance Metrics}}

\item{average}{logical. Should the rowwise distances be averaged as opposed
to summed?}

\item{...}{additional parameters to be passed to method function}
}
\description{
Given two alternative embeddings of a set of tokens or documents, \code{total_dist()}
computes a global metric of the distance between the alternatives (by default
the Wasserstein distance).
}
\details{
\code{total_dist()} computes the distance or similarity between the embeddings in
\code{x} and their direct counterparts in \code{y}. \code{average = TRUE} returns the
mean of these values, while \code{average = FALSE} (the default) returns the sum.

For more information on available methods, see \code{\link[=get_sims]{get_sims()}}.

\code{method = "euclidean"} and \code{average = FALSE} (the default for \code{total_dist()})
results in the \href{https://en.wikipedia.org/wiki/Wasserstein_metric}{Wasserstein distance},
if embeddings are taken as equally weighted point masses.

\code{average_sim()} is identical to \code{total_dist()}, but with different defaults.
}
