% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_embeddings.R
\name{write_embeddings}
\alias{write_embeddings}
\title{Save Embeddings in Fast-Loading Format}
\usage{
write_embeddings(x, file)
}
\arguments{
\item{x}{an embeddings object to save}

\item{file}{the name of a file or connection to save to}
}
\description{
Saves an embeddings object as a binary file that allows quick access to arbitrary rows
}
