% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_embedding.R
\name{average_embedding}
\alias{average_embedding}
\title{Average Embeddings}
\usage{
average_embedding(x, weights = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{an embeddings object}

\item{weights}{a numeric vector of weights for each row in x, \code{"trillion_word"}
(125,000 English word frequencies from \href{https://norvig.com/ngrams/count_1w.txt}{Peter Norvig's compilation},
derived from the Google Web Trillion Word Corpus), or smooth inverse frequencies
(SIF) calculated using the same list.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations?}
}
\description{
Calculate the (weighted) average of multiple embeddings.
}
\details{
For \code{weights = "trillion_word"}, tokens that do not appear in the word frequency
list are treated as if they appeared as often as the least frequent word in the list.
}
\examples{
happy_dict <- c("happy", "joy", "smile", "enjoy")
happy_dict_embeddings <- predict(glove_twitter_25d, happy_dict)

happy_dict_vec <- average_embedding(happy_dict_embeddings)
happy_dict_vec_weighted <- average_embedding(happy_dict_embeddings, weights = "trillion_word")
}
