% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{embeddings}
\alias{embeddings}
\alias{as.embeddings}
\alias{as.embeddings.default}
\alias{as.embeddings.data.frame}
\alias{is.embeddings}
\title{Embeddings Objects}
\usage{
embeddings(data = NA, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)

as.embeddings(x, ...)

\method{as.embeddings}{default}(x, ..., rowname_repair = TRUE, rebuild_token_index = TRUE)

\method{as.embeddings}{data.frame}(x, id_col = NULL, ..., rowname_repair = TRUE)

is.embeddings(x, ...)
}
\arguments{
\item{data}{an optional data vector (including a list or
    \code{\link[base]{expression}} vector).  Non-atomic classed \R objects are
    coerced by \code{\link[base]{as.vector}} and all attributes discarded.}

\item{nrow}{the desired number of rows.}

\item{ncol}{the desired number of columns.}

\item{byrow}{logical. If \code{FALSE} (the default) the matrix is
    filled by columns, otherwise the matrix is filled by rows.}

\item{dimnames}{a \code{\link[base]{dimnames}} attribute for the matrix:
    \code{NULL} or a \code{list} of length 2 giving the row and column
    names respectively.  An empty list is treated as \code{NULL}, and a
    list of length one as row names.  The list can be named, and the
    list names will be used as names for the dimensions.}

\item{x}{A data frame to be converted into embeddings.}

\item{...}{Additional arguments passed to or from other methods.}

\item{rowname_repair}{logical. If \code{TRUE} (the default), check that unique
rownames are provided, and name rows "doc_1", "doc_2", etc. if not.}

\item{rebuild_token_index}{logical. If \code{TRUE}, the hash table index will be rebuilt
even when \code{x} is an embeddings object.}

\item{id_col}{Optional name of a column to take row names from.}
}
\description{
An embeddings object is a numeric matrix with fast indexing by rownames
(generally tokens).
}
\details{
Fast row indexing is implemented using hash tables in native R environments.
The "token_index" attribute of an embeddings object stores the environment
that maps rownames to their corresponding indices.

If \code{dimnames} is not supplied, \code{embeddings} will automatically name rows
doc_1, doc_2, etc., and columns dim_1, dim_2, etc.
}
\examples{
random_mat <- matrix(
  sample(1:10, 20, replace = TRUE),
  nrow = 2,
  dimnames = list(c("happy", "sad"))
  )
random_embeddings <- as.embeddings(random_mat)
is.embeddings(random_embeddings[,2:5])

tibble::as_tibble(random_embeddings, rownames = "token")
}
