% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{as.embeddings}
\alias{as.embeddings}
\alias{as.embeddings.data.frame}
\alias{is.embeddings}
\title{Coercion and checking functions for embeddings objects}
\usage{
as.embeddings(x, ...)

\method{as.embeddings}{data.frame}(x, id_col = "token", ...)

is.embeddings(x, ...)
}
\arguments{
\item{x}{a numeric matrix or dataframe in which the only non-numeric column is specified by \code{id_col}}

\item{...}{additional arguments to be passed to class-specific methods}

\item{id_col}{column to take row names from. Defaults to "token"}
}
\description{
An embeddings object is a numeric matrix with rownames (generally tokens). Convert an
eligible input object into an embeddings object, or check whether an object
is an embeddings object.
}
\examples{
random_mat <- matrix(
  sample(1:10, 20, replace = TRUE),
  nrow = 2,
  dimnames = list(c("happy", "sad"))
  )
as.embeddings(random_mat)
}
