% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest.R
\name{find_nearest}
\alias{find_nearest}
\title{Find Nearest Tokens in Embedding Space}
\usage{
find_nearest(
  object,
  newdata,
  top_n = 10L,
  sim_func = cos_sim,
  each = FALSE,
  include_self = TRUE
)
}
\arguments{
\item{object}{an embeddings object made by \code{load_embeddings()} or \code{as.embeddings()}}

\item{newdata}{a character vector of tokens}

\item{top_n}{integer. How many nearest neighbors should be output? If \code{each = TRUE},
how many nearest neighbors should be output for each token?}

\item{sim_func}{a function that takes two vectors and outputs a scalar
similarity metric. Defaults to cosine similarity.}

\item{each}{logical. If \code{FALSE} (the default), the embeddings of newdata are
averaged and the function will output an embeddings object with the \code{top_n}
nearest tokens to the overall average embedding. If \code{TRUE}, the function will
output a named list with one embeddings object for each token in newdata.}

\item{include_self}{logical. Should the token(s) in \code{newdata} be included in
the output?}
}
\description{
Find Nearest Tokens in Embedding Space
}
\section{Value}{

An embeddings object or list of embeddings objects.
}

\examples{
   words <- c("happy", "sad")

   fake_mod <- as.embeddings(
    matrix(
      sample(1:10, 20, replace = TRUE),
      nrow = 2,
      dimnames = list(c("happy", "sad"))
      )
    )

   texts_embeddings <- predict(fake_mod, words)
   texts_embeddings
}
