% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embeddings.R
\name{load_embeddings}
\alias{load_embeddings}
\title{Load Embeddings Model From a Text or Binary File}
\usage{
load_embeddings(path)
}
\arguments{
\item{path}{a file path}
}
\description{
Reads GloVe text format, word2vec binary format, and most tabular formats (csv, tsv, etc.)
}
\details{
If using a custom tabular format, the file must have tokens in the first column and numbers in all other columns.
}
\section{Value}{

An embeddings object (a numeric matrix with tokens as rownames)
}

