% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embeddings.R
\name{load_embeddings}
\alias{load_embeddings}
\title{Load Pretrained GloVe, word2vec, and fastText Embeddings}
\usage{
load_embeddings(
  model,
  dir = NULL,
  words = NULL,
  save = TRUE,
  format = "original"
)
}
\arguments{
\item{model}{the name of a supported model}

\item{dir}{directory in which the model is or should be saved when \code{save = TRUE}.
The default is the working directory, \code{\link[=getwd]{getwd()}}. Dir can be set more
permanently using \code{options(embeddings.model.path = dir)}.}

\item{words}{optional list of words for which to retrieve embeddings.}

\item{save}{logical. Should the model be saved to \code{dir} if it does not
already exist there?}

\item{format}{the format in which the model should be saved if it does not
exist already. \code{"original"} (the default) saves the file as is.
Other options are \code{"csv"} or \code{"rds"}.}
}
\description{
Loads pretrained word embeddings. If the specified model has already been
downloaded, it is read from file with a call to \code{\link[=read_embeddings]{read_embeddings()}}. If not,
the model is retrieved from online sources and, by default, saved.
}
\details{
The following are supported models for download. Note that some models are very large.
If you know in advance which word embeddings you will need (e.g. the set of unique
tokens in your corpus), consider specifying this with the \code{words} parameter to save
memory and processing time.
\subsection{GloVe}{

\itemize{
\item \code{glove.42B.300d}: Common Crawl (42B tokens, 1.9M vocab, uncased, 300d). Downloaded from https://huggingface.co/stanfordnlp/glove.
This file is a zip archive and must temporarily be downloaded in its entirety even when \code{words} is specified.
\item \code{glove.840B.300d}: Common Crawl (840B tokens, 2.2M vocab, cased, 300d). Downloaded from https://huggingface.co/stanfordnlp/glove.
This file is a zip archive and must temporarily be downloaded in its entirety even when \code{words} is specified.
\item \code{glove.6B.50d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 50d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.6B.100d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 100d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.6B.200d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 200d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.6B.300d}: Wikipedia 2014 + Gigaword 5 (6B tokens, 400K vocab, uncased, 300d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.twitter.27B.25d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 25d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.twitter.27B.50d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 50d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.twitter.27B.100d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 100d). Downloaded from https://github.com/piskvorky/gensim-data
\item \code{glove.twitter.27B.200d}: Twitter (2B tweets, 27B tokens, 1.2M vocab, uncased, 200d). Downloaded from https://github.com/piskvorky/gensim-data
}
}

\subsection{word2vec}{

Note that reading word2vec bin files may be slower than other formats. If
read time is a concern, consider setting \code{format = "csv"} or \code{format = "rds"}.
\itemize{
\item \code{GoogleNews.vectors.negative300}: Trained with skip-gram on Google News (~100B tokens, 3M vocab, cased, 300d). Downloaded from https://github.com/piskvorky/gensim-data
}
}

\subsection{HistWords}{

By-decade skip-gram embeddings trained on Google Books and Corpus of Historical
American English (COHA). To specify a decade, replace \verb{[decade]} with the first
year of that decade (e.g. \code{eng.all_sgns.1800}, \code{eng.all_sgns.1990}).
Originally downloaded from https://nlp.stanford.edu/projects/histwords/
\itemize{
\item \code{coha.word_sgns.[decade]}: Trained on Corpus of Historical American English (Genre-Balanced American English, 1830s-2000s).
\item \code{coha.lemma_sgns.[decade]}: Same as above, with lemmas rather than words.
\item \code{eng.all_sgns.[decade]}: Trained on Google N-Grams eng-all (All English, 1800s-1990s).
\item \code{eng.fiction.all_sgns.[decade]}: Trained on Google N-Grams eng-fiction-all (English Fiction, 1800s-1990s).
\item \code{chi.sim.all_sgns.[decade]}: Trained on Google N-Grams chi-sim-all (Simplified Chinese, 1950-1990s).
\item \code{fre.all_sgns.[decade]}: Trained on Google N-Grams fre-all (French, 1800s-1990s).
\item \code{ger.all_sgns.[decade]}: Trained on Google N-Grams ger-all (German, 1800s-1990s).
}
}

\subsection{ConceptNet Numberbatch}{

Multilingual word embeddings trained using an ensemble that combines data from
word2vec, GloVe, OpenSubtitles, and the ConceptNet common sense knowledge database.
Tokens are prefixed with language codes. For example, the English word "token" is
labeled "/c/en/token".
Downloaded from https://github.com/commonsense/conceptnet-numberbatch
\itemize{
\item \code{numberbatch.19.08}: Multilingual (9.2M vocab, uncased, 300d)
}
}

\subsection{fastText}{

Downloaded from https://fasttext.cc/docs/en/crawl-vectors.html
\itemize{
\item \verb{cc.[language].300}: 300-dimensional word vectors for 157 languages,
trained with CBOW on Common Crawl and Wikipedia. To specify a language,
replace \verb{[language]} with one of the following language codes:
\code{af} (Afrikaans), \code{sq} (Albanian), \code{als} (Alemannic), \code{am} (Amharic), \code{ar}
(Arabic), \code{an} (Aragonese), \code{hy} (Armenian), \code{as} (Assamese), \code{ast}
(Asturian), \code{az} (Azerbaijani), \code{ba} (Bashkir), \code{eu} (Basque), \code{bar}
(Bavarian), \code{be} (Belarusian), \code{bn} (Bengali), \code{bh} (Bihari), \code{bpy}
(Bishnupriya Manipuri), \code{bs} (Bosnian), \code{br} (Breton), \code{bg} (Bulgarian),
\code{my} (Burmese), \code{ca} (Catalan), \code{ceb} (Cebuano), \code{bcl} (Central Bicolano),
\code{ce} (Chechen), \code{zh} (Chinese), \code{cv} (Chuvash), \code{co} (Corsican), \code{hr}
(Croatian), \code{cs} (Czech), \code{da} (Danish), \code{dv} (Divehi), \code{nl} (Dutch), \code{pa}
(Eastern Punjabi), \code{arz} (Egyptian Arabic), \code{eml} (Emilian-Romagnol), \code{en}
(English), \code{myv} (Erzya), \code{eo} (Esperanto), \code{et} (Estonian), \code{hif}
(Fiji Hindi), \code{fi} (Finnish), \code{fr} (French), \code{gl} (Galician), \code{ka}
(Georgian), \code{de} (German), \code{gom} (Goan Konkani), \code{el} (Greek), \code{gu}
(Gujarati), \code{ht} (Haitian), \code{he} (Hebrew), \code{mrj} (Hill Mari), \code{hi} (Hindi),
\code{hu} (Hungarian), \code{is} (Icelandic), \code{io} (Ido), \code{ilo} (Ilokano), \code{id}
(Indonesian), \code{ai} (Interlingua), \code{ga} (Irish), \code{it} (Italian), \code{ja}
(Japanese), \code{jv} (Javanese), \code{kn} (Kannada), \code{pam} (Kapampangan), \code{kk}
(Kazakh), \code{km} (Khmer), \code{ky} (Kirghiz), \code{ko} (Korean), \code{ku}
(Kurdish: Kurmanji), \code{ckb} (Kurdish: Sorani), \code{la} (Latin), \code{lv} (Latvian),
\code{li} (Limburgish), \code{lt} (Lithuanian), \code{lmo} (Lombard), \code{nds} (Low Saxon),
\code{lb} (Luxembourgish), \code{mk} (Macedonian), \code{mai} (Maithili), \code{mg} (Malagasy),
\code{ms} (Malay), \code{ml} (Malayalam), \code{mt} (Maltese), \code{gv} (Manx), \code{mr} (Marathi),
\code{mzn} (Mazandarani), \code{mhr} (Meadow Mari), \code{min} (Minangkabau), \code{xmf}
(Mingrelian), \code{mwl} (Mirandese), \code{mn} (Mongolian), \code{nah} (Nahuatl), \code{nap}
(Neapolitan), \code{ne} (Nepali), \code{new} (Newar), \code{frr} (North Frisian), \code{nso}
(Northern Sotho), \code{no} (Norwegian: Bokmål), \code{nn} (Norwegian: Nynorsk), \code{oc}
(Occitan), \code{or} (Oriya), \code{os} (Ossetian), \code{pfl} (Palatinate German), \code{ps}
(Pashto), \code{fa} (Persian), \code{pms} (Piedmontese), \code{pl} (Polish), \code{pt}
(Portuguese), \code{qu} (Quechua), \code{ro} (Romanian), \code{rm} (Romansh), \code{ru}
(Russian), \code{sah} (Sakha), \code{sa} (Sanskrit), \code{sc} (Sardinian), \code{sco} (Scots),
\code{gd} (Scottish Gaelic), \code{sr} (Serbian), \code{sh} (Serbo-Croatian), \code{scn}
(Sicilian), \code{sd} (Sindhi), \code{si} (Sinhalese), \code{sk} (Slovak), \code{sl}
(Slovenian), \code{so} (Somali), \code{azb} (Southern Azerbaijani), \code{es} (Spanish),
\code{su} (Sundanese), \code{sw} (Swahili), \code{sv} (Swedish), \code{tl} (Tagalog), \code{tg}
(Tajik), \code{ta} (Tamil), \code{tt} (Tatar), \code{te} (Telugu), \code{th} (Thai), \code{bo}
(Tibetan), \code{tr} (Turkish), \code{tk} (Turkmen), \code{uk} (Ukrainian), \code{hsb}
(Upper Sorbian), \code{ur} (Urdu), \code{ug} (Uyghur), \code{uz} (Uzbek), \code{vec} (Venetian),
\code{vi} (Vietnamese), \code{vo} (Volapük), \code{wa} (Walloon), \code{war} (Waray), \code{cy}
(Welsh), \code{vls} (West Flemish), \code{fy} (West Frisian), \code{pnb} (Western Punjabi),
\code{yi} (Yiddish), \code{yu} (Yoruba), \code{diq} (Zazaki), \code{zea} (Zeelandic)
}
}
}
\section{Value}{

An embeddings object (a numeric matrix with tokens as rownames)
}

\references{
Bojanowski, P., Grave, E., Joulin, A., and Mikolov, T. (2016). Enriching Word Vectors with Subword Information. arXiv preprint. https://arxiv.org/abs/1607.04606

Hamilton, W. L., Leskovec, J., and Jurafsky, D. (2016). Diachronic Word Embeddings Reveal Statistical Laws of Semantic Change. In Proceedings of the 54th Annual Meeting of the ACL. https://aclanthology.org/P16-1141/

Mikolov, T., Chen, K., Corrado, G., and Dean, J. (2013). Efficient Estimation of Word Representations in Vector Space. In Proceedings of Workshop at ICLR. https://arxiv.org/pdf/1301.3781

Pennington, J., Socher, R., and Manning, C. D. (2014). GloVe: Global Vectors for Word Representation. https://nlp.stanford.edu/projects/glove/

Speer, R., Chin, J., and Havasi, C. (2017). ConceptNet 5.5: An Open Multilingual Graph of General Knowledge. In proceedings of AAAI 2017. http://aaai.org/ocs/index.php/AAAI/AAAI17/paper/view/14972
}
