% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude}
\alias{magnitude}
\alias{magnitude.numeric}
\title{Magnitude of Embeddings}
\usage{
magnitude(x, ...)

\method{magnitude}{numeric}(x, ...)
}
\arguments{
\item{x}{a numeric vector, embeddings object, or list of numeric or embeddings objects}

\item{...}{additional parameters to be passed to methods}
}
\description{
The magnitude, norm, or length of a vector is its Euclidean distance from the origin.
}
\section{Value}{

a numeric vector with one item per embedding in \code{x}. If \code{x} is a list, the
function will be called recursively and return a list.
}

\examples{
vec <- c(1, 4, 2)
magnitude(vec)

valence_embeddings <- predict(glove_twitter_25d, c("good", "bad"))
magnitude(valence_embeddings)

embeddings_list <- find_nearest(glove_twitter_25d, c("good", "bad"), each = TRUE)
magnitude(embeddings_list)
}
