% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.embeddings.R
\name{predict.embeddings}
\alias{predict.embeddings}
\title{Retrieve Token Embeddings}
\usage{
\method{predict}{embeddings}(object, newdata)
}
\arguments{
\item{object}{an embeddings object made by \code{load_embeddings()} or \code{as.embeddings()}}

\item{newdata}{a character vector of tokens}
}
\description{
Retrieve Token Embeddings
}
\section{Value}{

An embeddings object with a row for each item in \code{newdata}
}

\examples{
words <- c("happy", "sad")

fake_mod <- as.embeddings(
    matrix(
        sample(1:10, 20, replace = TRUE),
        nrow = 2,
        dimnames = list(c("happy", "sad"))
    )
)

texts_embeddings <- predict(fake_mod, words)
texts_embeddings
}
