% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings-predict.R
\name{predict.embeddings}
\alias{predict.embeddings}
\title{Retrieve Token Embeddings}
\usage{
\method{predict}{embeddings}(object, newdata, .keep_missing = FALSE)
}
\arguments{
\item{object}{an embeddings object made by \code{load_embeddings()} or \code{as.embeddings()}}

\item{newdata}{a character vector of tokens}

\item{.keep_missing}{logical. What should be done about items in \code{newdata}
that are not present in the embeddings object? If \code{FALSE} (the default), they
will be ignored. If \code{TRUE}, they will be returned as \code{NA}.}
}
\description{
Retrieve Token Embeddings
}
\section{Value}{

Either an embeddings object with a row for each item in \code{newdata}, or, when
\code{newdata} is of length 1, a named numeric vector.
}

\examples{
words <- c("happy", "sad")

texts_embeddings <- predict(glove_twitter_25d, words)
texts_embeddings
}
