% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utilities.R
\name{project_points_onto_line}
\alias{project_points_onto_line}
\title{Additional Vector Utilities}
\usage{
project_points_onto_line(line_start, line_end, points_df)
}
\arguments{
\item{line_start}{a vector of representing the coordinates of the line's start}

\item{line_end}{a vector representing the coordinates of the line's end}

\item{points_df}{a dataframe in which each row represents the coordinates of a
point to be projected onto the line}
}
\description{
Additional Vector Utilities
}
\details{
`project_points_onto_line` calculated the projections of a group of points onto
a line defined by two end points. It is useful for graphing the positions of
embeddings with respect to an anchored vector.
}
