% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensionality.R
\name{reduce_dimensionality}
\alias{reduce_dimensionality}
\alias{reduce_dimensionality.data.frame}
\alias{reduce_dimensionality.embeddings}
\alias{reduce_dimensionality.matrix}
\title{Reduce the Dimensionality of Embeddings}
\usage{
reduce_dimensionality(x, ...)

\method{reduce_dimensionality}{data.frame}(x, cols, reduce_to, scale = FALSE, ...)

\method{reduce_dimensionality}{embeddings}(x, reduce_to, scale = FALSE, ...)

\method{reduce_dimensionality}{matrix}(x, reduce_to, scale = FALSE, ...)
}
\arguments{
\item{x}{a dataframe with one embedding per row}

\item{...}{additional parameters to be passed to class-specific methods}

\item{cols}{tidyselect - columns that contain numeric embedding values}

\item{reduce_to}{number of dimensions to keep}

\item{scale}{perform scaling in addition to centering?}
}
\description{
Includes methods for dataframes (in the style of \code{dplyr}) embeddings objects, and matrices.
}
\details{
The file must have tokens in the first column and numbers in all other columns.
}
\section{Value}{

An embeddings object (a numeric matrix with tokens as rownames)
}

