% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensionality.R
\name{reduce_dimensionality}
\alias{reduce_dimensionality}
\alias{reduce_dimensionality.data.frame}
\alias{reduce_dimensionality.embeddings}
\title{Reduce the Dimensionality of Embeddings}
\usage{
reduce_dimensionality(x, ...)

\method{reduce_dimensionality}{data.frame}(x, cols, reduce_to, center = TRUE, scale = FALSE, ...)

\method{reduce_dimensionality}{embeddings}(x, reduce_to, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{x}{a dataframe, embeddings object, or matrix with one embedding per row}

\item{...}{additional parameters to be passed to class-specific methods}

\item{cols}{tidyselect - columns that contain numeric embedding values}

\item{reduce_to}{number of dimensions to keep.
The value is passed to \code{\link[stats:prcomp]{stats::prcomp()}} as \code{rank.}.}

\item{center}{logical. Should dimensions be shifted to be centered at zero?}

\item{scale}{logical. Should dimensions be scaled to have unit variance?}
}
\description{
Includes methods for dataframes (in the style of \code{dplyr}), embeddings objects, and matrices.
}
\details{
By default, \code{reduce_dimensionality()}, performs principle components analysis
(PCA) without column normalization, and outputs the rotated data. If \code{center = FALSE}
and \code{scale = FALSE}, this is equivalent to singular value decomposition (SVD),
\eqn{X = U \Sigma V^{T}}, where the output columns are equal to the first
\code{reduce_to} columns of \eqn{U \Sigma}.
}
\section{Value}{

An object of the same class as x, with the same number of rows but fewer columns.
Reduced columns in the output will be named "PC1", "PC2", etc.
}

\examples{
glove_2d <- reduce_dimensionality(glove_twitter_25d, 2)
glove_2d

library(tibble)
glove_tbl <- as_tibble(glove_twitter_25d, rownames = "token")
glove_tbl_2d <- glove_tbl |> reduce_dimensionality(dim_1:dim_25, 2)
glove_tbl_2d
}
