% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_embedding.R
\name{textstat_embedding}
\alias{textstat_embedding}
\title{Get Text Embeddings by Averaging Word Embeddings}
\usage{
textstat_embedding(dfm, model)
}
\arguments{
\item{dfm}{a quanteda dfm}

\item{model}{an embeddings object made by `load_embeddings()` or `as.embeddings()`}
}
\description{
Get Text Embeddings by Averaging Word Embeddings
}
\section{Value}{

A tibble with columns `doc_id`, and `V1`, `V2`, etc.
}

\examples{
   texts <- c("this says one thing", "and this says another")
   texts_dfm <- dfm(tokens(texts))

   fake_mod <- as.embeddings(
    matrix(
      sample(1:10, 30, replace = TRUE),
      nrow = 6,
      dimnames = list(c("this", "and", "says", "one", "thing", "another"))
      )
    )

   texts_embeddings <- textstat_embeddings(texts_dfm, fake_mod)
   texts_embeddings
}
